Revert "Move consent or pay functionality to shared code"
This reverts commit 77aab7eccf3a42ab0fb1819f4ecf57b6558e3e5b.
