+++ libs/@guardian/libs/src/consent-management-platform/isConsentOrPay.ts
@@ -1,13 +1,10 @@
-import type { ConsentOrPayCurrency } from '../countries/consentOrPay';
-import {
-	consentOrPayCurrencyMap,
-	isConsentOrPayCountry,
-} from '../countries/consentOrPay';
 import type { CountryCode } from '../index.test';
 import { isGuardianDomain } from './lib/domain';
-
-// Re-export for backward compatibility
-export { isConsentOrPayCountry };
+import {
+	consentOrPayCountries,
+	consentOrPayCurrencyMap,
+} from './lib/sourcepointConfig';
+import type { Currency } from './lib/sourcepointConfig';
 
 let _isConsentOrPay = false;
 
@@ -19,9 +16,11 @@ export const getIsConsentOrPay = (): boolean => {
 	return _isConsentOrPay;
 };
 
-export const getConsentOrPayCurrency = (
-	countryCode: CountryCode,
-): ConsentOrPayCurrency => {
+export const isConsentOrPayCountry = (countryCode: CountryCode) => {
+	return consentOrPayCountries.includes(countryCode);
+};
+
+export const getConsentOrPayCurrency = (countryCode: CountryCode): Currency => {
 	if (!isConsentOrPayCountry(countryCode)) {
 		return 'GBP';
 	}
