+++ libs/@guardian/libs/src/consent-management-platform/lib/sourcepointConfig.ts
@@ -26,3 +26,58 @@ export const SourcePointChoiceTypes = {
 	RejectAll: 13,
 	Dismiss: 15,
 } as const;
+
+export const consentOrPayCountries = [
+	'GB',
+	'JE',
+	'GG',
+	'IM',
+	'GI',
+	'AT',
+	'BE',
+	'BG',
+	'HR',
+	'CY',
+	'CZ',
+	'DK',
+	'EE',
+	'FI',
+	'FR',
+	'DE',
+	'GR',
+	'HU',
+	'IS',
+	'IE',
+	'IT',
+	'LV',
+	'LT',
+	'LU',
+	'MT',
+	'NL',
+	'NO',
+	'PL',
+	'PT',
+	'RO',
+	'SK',
+	'SI',
+	'ES',
+	'SE',
+	'CH',
+];
+
+export type Currency = 'GBP' | 'EUR';
+
+// Return EUR unless otherwise specified (to avoid duplicating the list of European countries)
+export const consentOrPayCurrencyMap = new Proxy(
+	{
+		GB: 'GBP' as Currency,
+	} as Record<string, Currency>,
+	{
+		get(target, prop): Currency {
+			if (typeof prop === 'string') {
+				return target[prop] ?? 'EUR';
+			}
+			return 'EUR';
+		},
+	},
+);
