+++ libs/@guardian/libs/src/consent-management-platform/types/window.d.ts
@@ -1,7 +1,6 @@
-import type { ConsentOrPayCurrency } from '../../countries/consentOrPay';
 import type { getConsentFor } from '../getConsentFor';
 import type { Property } from '../lib/property';
-import type { EndPoint } from '../lib/sourcepointConfig';
+import type { Currency, EndPoint } from '../lib/sourcepointConfig';
 import type { onConsent } from '../onConsent';
 import type { onConsentChange } from '../onConsentChange';
 import type { AUSData } from './aus';
@@ -53,7 +52,7 @@ declare global {
 						excludePage: boolean;
 						isCorP: boolean;
 						isUserSignedIn: boolean;
-						corPCurrency: ConsentOrPayCurrency;
+						corPCurrency: Currency;
 					};
 				};
 				usnat?: {
