+++ libs/@guardian/libs/src/index.test.ts
@@ -10,8 +10,6 @@ describe('The package', () => {
 			'ArticleSpecial',
 			'Pillar',
 			'cmp',
-			'consentOrPayCountries',
-			'consentOrPayCurrencyMap',
 			'countries',
 			'debug',
 			'getConsentFor',
@@ -21,7 +19,6 @@ describe('The package', () => {
 			'getMeasures',
 			'getSwitches',
 			'isBoolean',
-			'isConsentOrPayCountry',
 			'isNonNullable',
 			'isObject',
 			'isOneOf',
@@ -49,7 +46,6 @@ export type {
 	ArticleTheme,
 	Country,
 	CountryCode,
-	ConsentOrPayCurrency,
 	OphanABEvent,
 	OphanABPayload,
 	OphanABTestMeta,
