+++ libs/@guardian/libs/src/index.ts
@@ -32,12 +32,6 @@ export type { Country } from './countries/@types/Country';
 export type { CountryCode } from './countries/@types/CountryCode';
 export { countries } from './countries/countries';
 export { getCountryByCountryCode } from './countries/getCountryByCountryCode';
-export type { ConsentOrPayCurrency } from './countries/consentOrPay';
-export {
-	consentOrPayCountries,
-	consentOrPayCurrencyMap,
-	isConsentOrPayCountry,
-} from './countries/consentOrPay';
 
 export { timeAgo } from './datetime/timeAgo';
